add skip list CmakeList

+++ string/kmp.hpp
@@ -5,13 +5,13 @@
 class KMPAlgorithm
 {
 public:
-	static int32_t get_index_of(const std::string& str1, const std::string& str2)
+	static size_t get_index_of(const std::string& str1, const std::string& str2)
 	{
 		std::size_t s1 = str1.length();
 		std::size_t s2 = str2.length();
 		if (s1 == 0 || s2 == 0 || s1 < s2) return -1;
 
-		std::vector<int32_t> str2_next_array = get_next_array(str2);
+		std::vector<size_t> str2_next_array = get_next_array(str2);
 		size_t i = 0;
 		size_t j = 0;
 		while (i < s1 && j < s2)
@@ -40,16 +40,16 @@ private:
 	/// </summary>
 	/// <param name="str"></param>
 	/// <returns></returns>
-	static std::vector<int32_t> get_next_array(const std::string& str)
+	static std::vector<size_t> get_next_array(const std::string& str)
 	{
-		std::vector<int32_t> next;
+		std::vector<size_t> next;
 		next.resize(str.length());
 		next.at(0) = -1; //第一个字符规定为-1
 		if (str.length() == 1) return next;
 
 		next.at(1) = 0; //第二个字符规定为0
 		size_t idx = 2;
-		int32_t cn = 0;
+		size_t cn = 0;
 		while (idx < str.length())
 		{
 			if (str.at(cn) == str.at(idx - 1))
